+++ src/Channel/Channel.cpp
@@ -3,16 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   Channel.cpp                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fbily <fbily@student.42.fr>                +#+  +:+       +#+        */
+/*   By: zhamdouc <zhamdouc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 17:28:51 by fbily             #+#    #+#             */
-/*   Updated: 2023/09/15 13:08:20 by fbily            ###   ########.fr       */
+/*   Updated: 2023/09/15 14:45:29 by zhamdouc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/Channel.hpp"
-
-/******		Constructors	******/
 
 Channel::Channel() {}
 
@@ -38,11 +36,7 @@ Channel::Channel(std::string name, std::string key, User *user) : _name(name), _
 	}
 }
 
-/******		Destructor	******/
-
 Channel::~Channel() {}
-
-/******		Operator	******/
 
 Channel& Channel::operator=(const Channel& rhs)
 {
@@ -50,94 +44,15 @@ Channel& Channel::operator=(const Channel& rhs)
 	return (*this);
 }
 
-/******		Member functions	******/
-
-std::string	Channel::getName() const
-{
-	return (this->_name);
-}
-
-std::string Channel::getTopic() const
-{
-	return (this->_topic);
-}
-
-bool Channel::getModeTopic() const
-{
-	return (this->_topicMode);
-}
-
-std::string	Channel::getTopicDate() const
-{
-	return (this->_topicTime);
-}
-
-std::string	Channel::getTopicUser() const
-{
-	return (this->_topicUser);
-}
-
-std::string Channel::getCreationDate() const
-{
-	return (this->_creationDate);
-}
-
-std::string Channel::getModestring() const
-{
-	std::string	modestring = "+";
-
-	if (this->_inviteMode)
-		modestring.append("i");
-	if (this->_topicMode)
-		modestring.append("t");
-	if (this->_isKey == true)
-		modestring.append("k");
-	if (this->_maxUsers != 0)
-	{
-		std::ostringstream oss;
-		oss << this->_maxUsers;
-		modestring.append("l ");
-		modestring.append(oss.str());
-	}
-	if (modestring == "+")
-		modestring.clear();
-	return (modestring);
-}
-
-unsigned int Channel::getNbUser() const
-{
-	return (this->_nbUser);
-}
-
-void Channel::setNbUser(int nb)
-{
-	this->_nbUser += nb;
-}
-
-void Channel::setTopic(User *user, std::string newTopic)
-{
-	time_t	now = time(0);
-	std::ostringstream oss;
-	oss << now;
-	if(newTopic == ":")
-		this->_topic = "";
-	else
-		this->_topic = newTopic;
-	this->_topicTime = oss.str(); // retiens l'heure de modif
-	this->_topicUser = user->getNickname(); // retien qui l'a modif
-	sendMessage(user, RPL_TOPIC(user->getNickname(), user->getServer(), this->_name, this->_topic));
-	sendMessage(user, RPL_TOPICWHOTIME(user->getNickname(), user->getServer(), this->_name, this->_topicUser, this->_topicTime));
-}
-
 std::string	Channel::joinChannel(std::string key, User *user)
 {
 	if (_inviteMode ==  true)
 	{
 		if (foundInvited(user->getNickname()) == false)
-			return (ERR_INVITEONLYCHAN(user->getNickname(), this->getName()));//il est pas invite ce batard  
+			return (ERR_INVITEONLYCHAN(user->getNickname(), this->getName()));  
 	}
 	if (this->_maxUsers != 0 && this->_nbUser >= this->_maxUsers)
-		return (ERR_CHANNELISFULL(user->getNickname(), this->getName()));//erreur plus de place
+		return (ERR_CHANNELISFULL(user->getNickname(), this->getName()));
 	
 	if (this->_isKey == false || (this->_isKey == true && key == this->_key))
 	{
@@ -146,7 +61,7 @@ std::string	Channel::joinChannel(std::string key, User *user)
 		return (':' + user->getNickname() + "!" + user->getNickname() + "@" + user->getHostname() + " JOIN " + this->getName());
 	}
 	else
-		return (ERR_BADCHANNELKEY(user->getNickname(), this->getName()));//erreur de mot de passe 
+		return (ERR_BADCHANNELKEY(user->getNickname(), this->getName()));
 }
 
 bool Channel::foundUser(std::string nickname)
@@ -157,7 +72,7 @@ bool Channel::foundUser(std::string nickname)
 		if (nickname != (*it)->getNickname())
 			it++;
 		else
-			return (true);//trouve
+			return (true);
 	}
 	return (false);
 }
@@ -266,93 +181,6 @@ void	Channel::whoList(User *user)
 	}
 }
 
-bool Channel::hasDuplicates(const std::string& str) 
-{
-	if(str.size() > 6 )
-		return (false);
-    for (std::size_t i = 0; i < str.size(); i++) 
-	{
-        for (std::size_t j = i + 1; j < str.size(); j++) 
-		{
-            if (str[i] == str[j])
-                return true;
-        }
-    }
-    return false;
-}
-
-void	Channel::set_inviteMode(User *user, int sign)
-{
-	if (sign == 1)
-	{
-		this->_inviteMode = true;
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " +i");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " +i");
-	}
-	if (sign == -1)
-	{
-		this->_inviteMode = false;
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " -i");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " -i");
-	}
-}
-
-void	Channel::set_topicMode(User *user, int sign)
-{
-	if (sign == 1)
-	{
-		this->_topicMode = true;
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " +t");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " +t");
-	}
-	if (sign == -1)
-	{
-		this->_topicMode = false;
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " -t");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " -t");
-	}
-}
-
-void	Channel::set_key(User *user, std::string argument, int pos_argument, int sign)
-{
-	if (sign == -1)
-	{
-		this->_isKey = false;
-		this->_key = "";
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " -k");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " -k");
-		return ;
-	}
-	std::string arg_1;
-	int pos_bis = 0;
-	size_t pos;
-	
-	while (!argument.empty() && pos_bis <= pos_argument)
-	{
-		pos = argument.find(' ');
-		arg_1 = argument.substr(0, argument.find(' '));
-		if (pos_bis == pos_argument)
-			break;
-		if (pos == std::string::npos)
-		{
-			argument.erase(0, arg_1.length());  
-			break;
-		}
-		else
-			argument.erase(0, arg_1.length() + 1);// +1 pour supprimer l'espace suivant
-		pos_bis++;
-	}
-	if (!argument.empty())
-	{
-		if (arg_1 == "x")
-			return (user->sendReply(":" + user->getServer() + " 525 " + user->getNickname() + ' ' + this->_name + " :Key is not well-formed (x is not a valid key)"));
-		this->_key = true;
-		this->_key = arg_1;
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " +k");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " +k");
-	}
-}
-
 void	Channel::kickModeOperator(std::string target)
 {
 	std::vector<User *>::iterator	it = this->_Operators.begin();
@@ -395,74 +223,6 @@ void	Channel::kickModeInvited(std::string target)
 	}
 }
 
-User* Channel::returnOperator(std::string nickname)
-{
-	std::vector<User *>::iterator it = this->_Operators.begin();
-	while (it != this->_Operators.end())
-	{
-		if (nickname != (*it)->getNickname())
-			it++;
-		else
-			return (*it);
-	}
-	return (*it);
-}
-
-User* Channel::returnUser(std::string nickname)
-{
-	std::vector<User *>::iterator it = this->_Users.begin();
-	while (it != this->_Users.end())
-	{
-		if (nickname != (*it)->getNickname())
-			it++;
-		else
-			return (*it);
-	}
-	return (*it);
-}
-
-void	Channel::set_op(User *user, std::string argument, int pos_argument, int sign)
-{
-	std::string arg_1;
-	int pos_bis = 0;
-	size_t pos = 0;
-
-	while (!argument.empty() && pos_bis <= pos_argument)
-	{
-		pos = argument.find(' ');
-		arg_1 = argument.substr(0, argument.find(' '));
-		if (pos_bis == pos_argument)
-			break;
-		if (pos == std::string::npos)
-		{
-			argument.erase(0, arg_1.length());  
-			break;
-		}
-		else
-			argument.erase(0, arg_1.length() + 1);// +1 pour supprimer l'espace suivant
-		pos_bis++;
-	}
-	if (foundOperator(user->getNickname()))
-	{
-		if (sign == -1 && foundOperator(arg_1) == true)
-		{
-			this->_Users.push_back(returnOperator(arg_1));
-			kickModeOperator(arg_1);
-			user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " -o " + arg_1);
-			sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " -o " + arg_1);
-		}
-		if (sign == 1 && foundOperator(arg_1) == false)
-		{
-			this->_Operators.push_back(returnUser(arg_1));
-			kickModeUser(arg_1);
-			user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " +o " + arg_1);
-			sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " +o " + arg_1);
-		}
-	}
-	else
-		return ; //Un user ne peut pas retirer de privilege
-}
-
 // bool safeStringToInt(const std::string& str, int& result)
 // {
 //     std::stringstream ss(str);
@@ -485,74 +245,13 @@ void	Channel::set_op(User *user, std::string argument, int pos_argument, int sig
 //     return true;
 // }
 
-unsigned int	Channel::getMaxUsers() const
-{
-	return (this->_maxUsers);
-}
-
-void	Channel::set_maxUsers(User *user, std::string argument, int pos_argument, int sign)
-{
-	if (sign == -1)
-	{
-		this->_maxUsers = 0;
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " -l");
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " -l");
-		return ;
-	}
-	//checker quelle est le max pour un channel
-	std::string arg_1;//faire le parsing du int 
-	size_t pos = 0;
-	int pos_bis = 0;
-	
-	while (!argument.empty() && pos_bis <= pos_argument)
-	{
-		pos = argument.find(' ');
-		arg_1 = argument.substr(0, argument.find(' '));
-		if (pos_bis == pos_argument)
-			break;
-		if (pos == std::string::npos)
-		{
-			argument.erase(0, arg_1.length());  
-			break;
-		}
-		else
-			argument.erase(0, arg_1.length() + 1);// +1 pour supprimer l'espace suivant
-		pos_bis++;
-	}
-	if (!argument.empty() && std::atoi(arg_1.c_str()) > static_cast<int>(this->_nbUser))
-	{
-		this->_maxUsers = std::atoi(arg_1.c_str());
-		user->sendReply(":" + user->getNickname() + " MODE " + this->_name + " +l " + arg_1);
-		sendMessage(user, ":" + user->getNickname() + " MODE " + this->_name + " +l " + arg_1);
-	}
-}
-
 void	Channel::modeChannel(User *user, std::string modestring, std::string argument)
 {
-	/* est ce que on doit envoyer un message si on essaye de mettre a vrai une valeur deja vrai
-	quelle erreur envoyer en cas de muavais modestring */
-	/* MODE - Changer le mode du channel :
-		— i : Définir/supprimer le canal sur invitation uniquement (pas besoin d'argument)
-		— t : Définir/supprimer les restrictions de la commande TOPIC pour les opérateurs de canaux (pas besoin d'argument)
-		()— k : Définir/supprimer la clé du canal (mot de passe) (besoin d'un argument)
-		()— o : Donner/retirer le privilège de l’opérateur de canal (besoin d'un argument)
-		()— l : Définir/supprimer la limite d’utilisateurs pour le canal ()*/
-	/* Pour l'instant je suis parti du principe que tu peux m'envoyer un seul operator + ou -, et tous les modes en meme temps sans doublon
-	ce qui fait donc au max 3 argument  */
 	int pos_argument = 0;
 	int sign = 0;
-	
-	//if (hasDuplicates(modestring))
-	//	return ;//erreur doublon
-	if(modestring[0] == '+' || modestring[0] == '-')//pour l'instant j'aurtorise que l'un des deux mais je peux sinon mettre en place la possibilite de swap entre les deux mais faudra changer la fonction qui check les duplicates
+
+	if(modestring[0] == '+' || modestring[0] == '-')
 	{
-		/* pour gerer le plus et le moins en meme temps, on peut juste rajouter (dans la boucle et avant):
-		int operator = 0;
-		if(modestring[i] == '+')
-			operator = 0
-		if(modestring[i] == '-')
-			operator =  1
-		puis envoyer operator au lieu de modestring dans les fonction*/
 		if(modestring[0] == '+')
 			sign = 1;
 		if(modestring[0] == '-')
